Extended main script to run PRE
The process rate estimator is run in the main script. The function to run the PRE allows to prevent non-negative solutions; however, this will possibly greatly reduce the number of found solutions, even to zero.

The PRE doesn't return all found solutions, but rather the quantiles that describe those. This prevents saving too much information.